giovanni - wk7 day3 - wk8_petition5 - pls matt help me fixit

+++ routes/petition/signed/index.js
@@ -5,19 +5,28 @@ const db = require( '../../../modules/dbQuery' );
 signed.route( '/' )
 
     .all( ( req, res, next ) => {
-        if ( req.session.signature_id == false ) {
-            res.redirect( '/petition' );
-        } else {
-            console.log('we are here yo!');
-            next();
+        if ( !req.session || !req.session.user_id ) {
+            res.render( 'error', {
+                message: 'user unauthorized'
+            } );
+            // res.redirect( '/' );
         }
+        if ( req.session.signature_id == false ) {
+            res.render( 'error', {
+                message: 'you have already signed the petition'
+            } );
+            // res.redirect( '/petition' );
+        }
+        next();
     } )
 
 
     .get( ( req, res ) => {
         // return the signature of the provided id
         const user_id = req.session.user_id;
-        let signedData = {};
+        let signedData = {
+            csrfToken: req.csrfToken()
+        };
 
         const getSignature = db.getSignature( user_id ).then( ( signature ) => {
             return signedData.signature = signature;
@@ -29,19 +38,28 @@ signed.route( '/' )
 
 
         Promise.all( [ getSignature, getSigners ] ).then( () => {
-            console.log('about to render /petiotion/signed');
+            console.log( 'about to render /petiotion/signed' );
             res.render( 'signed', signedData );
         } ).catch( ( err ) => {
             console.error( err.stack );
         } );
     } );
 
-signed.get( '/unsigning', ( req, res ) => {
-    db.deleteSignature( req.session.user_id ).then( () => {
-        req.session.signature_id = false;
-        res.redirect( '/petition' );
-    } );
+signed.get( '/tounsign', ( req, res ) => {
+    if ( req.session.user_id && req.session.signature_id ) {
+        console.log( 'delete signature request' );
+        db.deleteSignature( req.session.user_id ).then( () => {
+            req.session.signature_id = false;
+            res.redirect( '/petition' );
+        } );
+
+    } else {
+        res.render( 'error', {
+            message: 'user unauthorized'
+        } );
+    }
 } );
+
 
 /* MODULE EXPORTS */
 module.exports = signed;
